+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project are documented in this file.
 
 The format follows Keep a Changelog, and this project uses Semantic Versioning.
 
-## [1.3.3] - TBD
+## [1.3.3] - 2026-07-24
 
 ### Added